pyproject.toml
[project]
name = "zkt-access-gateway"
version = "1.0.0"
description = "Dockerized Python project bridging pyzkaccess via Wine"
authors = [
    { name = "Developer", email = "developer@example.com" }
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.85.0",
    "uvicorn>=0.18.0",
    "paho-mqtt>=1.6.1,<2.0.0",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.1",
    "pydantic>=2.4.0",
]
